improve: more concept, more debug, serializable wrapper

+++ etoile/gear/Object.cc
@@ -113,7 +113,7 @@ namespace etoile
       // dump the permissions.
       std::cout << alignment << elle::Dumpable::Shift << elle::Dumpable::Shift
                 << "[Permissions] "
-                << this->rights.permissions
+                << (int)this->rights.permissions
                 << std::endl;
 
       // dump the key.
